Change mapType to be deserialized to map[string]interface{}
This fixes #97 by setting the type of the keysof the deserialized maps to string.
Strings as keys are a safe default, since the JSONEncoder can handle them.

Please note that this is an api-breaking change for some applications, such as those
supporting only msgpack and checking explicitly against map[interface{}]interface{}
For JSON clients, no difference should be seen.