Handle busts of messages with send timeout instead of immediate drop
This change adds more progressive results examples and tests.  Testing found
that bursts of messages could sometimes lead to the router dropping messages
if a sufficient number arrived before they could be sent to the recipient.
This is fixed by waiting for a short timeout before dropping messages, giving
time for messages to be sent, or at least bufferred.