Provide client-side session filtering tool
This utility uses a client session to monitor session that join and leave the WAMP realm that the monitor client is connected to.  Named filters can be added to the monitor, where each filter is given a function that determines if a session is allowed and maintains a associated whitelist of session IDs.  The whitelist for a filter can be retrieved at any time, and is suitable for use as a whitelist when publishing a message.

A filter function may have whatever functionality is necessary for the client to determine whether or not each session is allowed.  This means complex filtering can be implemented without affecting the performance of the router, since the work is done by the clients that care to do filtering.