Fix blocking with slow clients
This fixes issue #159.  When a client leaves the realm or is disconnected by the router, do not attempt to send more messages to the client.

Web and raw socker server's outbound per-client channel size is configurable.

Progressive results use backpressure to limit influx of results when results are being produced faster than they are being consumed.