Do not decode msgpack raw bytes as string
When the router decodes a msgpack message that contains raw bytes, the router was decoding the bytes into a string.  When this data was then send out to destination client it was then encoded as a string and no longer as bytes.  This caused clients expecting bytes to get a string unexpectedly.

This is fixed by using the new msgpack spec and disabling the option to convert raw bytes to string.

This fixes issue #214