Provide ping handler that responds when client blocked
The websocket only handles responding to pings when reading messages.  If the client is blocked handling incoming messages and does not read additional messages for some time, websocker keep-alive pings are not responded to.  This may lead to the server dropping the websocket connection thinking that the client is dead.  This PR provides a ping handler that answers pings even when the client if not reading messages.

Addresses issue #245